Refactor to organize shims by Object
